refactor(evaluator): split engine/mod.rs into focused submodules
Issue #391 item 1: extract cohesive clusters from the oversized
src/evaluator/engine/mod.rs (1656 -> 1102 lines) into focused submodules,
pure code motion with no behavior change:

- output.rs: is_message_bearing, has_message_bearing_match
- pattern_recovery.rs: log_pattern_operand_skip, evaluate_children_or_warn
- subroutine.rs: SubroutineScope, evaluate_use_rule, USE_WITHOUT_RULE_ENV_WARNED
- value_eval.rs: evaluate_pattern_rule, evaluate_value_rule, string_ordering_display_value

mod.rs retains the dispatch loop. The ~700-line evaluate_rules is a single
GOTCHAS-pinned dispatch function that cannot be split without behavior-changing
helper extraction, so mod.rs stays over the 500-600 ceiling by design.

No behavior change: full suite green (1608), clippy/fmt clean both feature sets.
