+++ src/evaluator/engine/mod.rs
@@ -19,6 +19,16 @@ use super::{EvaluationContext, RecursionGuard, RuleMatch, offset, operators, typ
 use log::{debug, warn};
 use std::sync::atomic::{AtomicBool, Ordering};
 
+mod output;
+mod pattern_recovery;
+mod subroutine;
+mod value_eval;
+
+pub(crate) use output::*;
+pub(crate) use pattern_recovery::*;
+pub(crate) use subroutine::*;
+pub(crate) use value_eval::*;
+
 /// RAII guard that saves the GNU `file` previous-match anchor **and**
 /// `base_offset` on entry and restores both on drop.
 ///
@@ -66,68 +76,6 @@ impl Drop for AnchorScope<'_> {
         self.context.set_base_offset(self.saved_base);
     }
 }
-
-/// RAII guard for `MetaType::Use` subroutine dispatch.
-///
-/// Saves `last_match_end` and `base_offset` on entry, seeds the context
-/// with the use-site offset (for both fields so that a subroutine's
-/// `&0` relative offset resolves to the use-site and its positive
-/// absolute offsets bias against the use-site per magic(5)), and
-/// restores both on drop.
-///
-/// This is the safety net for early-return paths inside
-/// `evaluate_use_rule`: a `RecursionGuard::enter` failure or a
-/// `Timeout`/`RecursionLimitExceeded` inside the subroutine body would
-/// otherwise leave the caller's context with corrupted anchor and
-/// base-offset state. The guard's `Drop` impl restores both fields on
-/// every exit path, error or success.
-struct SubroutineScope<'a> {
-    context: &'a mut EvaluationContext,
-    saved_anchor: usize,
-    saved_base: usize,
-    saved_flip: bool,
-}
-
-impl<'a> SubroutineScope<'a> {
-    /// Enter a subroutine body. `flip_use` is the `\^` prefix on the
-    /// invoking `use` site; the effective flip inside the body is the
-    /// caller's flip state XOR `flip_use` -- matching libmagic's `flip =
-    /// !flip` toggle, which nests: a `\^use` inside an already-flipped
-    /// subroutine un-flips. Restored on drop along with anchor and base.
-    fn enter(context: &'a mut EvaluationContext, use_site: usize, flip_use: bool) -> Self {
-        let saved_anchor = context.last_match_end();
-        let saved_base = context.base_offset();
-        let saved_flip = context.flip_endian();
-        context.set_last_match_end(use_site);
-        context.set_base_offset(use_site);
-        context.set_flip_endian(saved_flip ^ flip_use);
-        Self {
-            context,
-            saved_anchor,
-            saved_base,
-            saved_flip,
-        }
-    }
-
-    fn context(&mut self) -> &mut EvaluationContext {
-        self.context
-    }
-}
-
-impl Drop for SubroutineScope<'_> {
-    fn drop(&mut self) {
-        self.context.set_last_match_end(self.saved_anchor);
-        self.context.set_base_offset(self.saved_base);
-        self.context.set_flip_endian(self.saved_flip);
-    }
-}
-
-/// Process-local once guard for the "use directive without rule environment"
-/// warning. Ensures we surface the misconfiguration exactly once per process
-/// so low-level programmatic consumers of [`evaluate_rules`] (tests, fuzz
-/// harnesses) that intentionally run without a `MagicDatabase`-attached
-/// environment do not flood the log on every `Use` rule they encounter.
-static USE_WITHOUT_RULE_ENV_WARNED: AtomicBool = AtomicBool::new(false);
 
 /// Process-local once guard for the "`evaluate_rules_with_config` called
 /// with an `indirect` rule but without a `RuleEnvironment`" warning.
@@ -300,508 +248,6 @@ fn evaluate_single_rule_with_anchor(
         )?,
     };
     Ok(matched.then_some((absolute_offset, read_value)))
-}
-
-/// Evaluate a `TypeKind::Meta(MetaType::Use { name, .. })` rule inline.
-///
-/// Looks up `name` in the context's rule environment, temporarily sets the
-/// GNU `file` previous-match anchor to the resolved offset, and recursively
-/// evaluates the subroutine's rules against `buffer`. Any matches produced
-/// by the subroutine are returned in document order and are intended to be
-/// pushed into the caller's match vector *before* the synthetic `Use` match
-/// itself (matching GNU `file` behavior where a `use` site is replaced by
-/// its expansion in the output).
-///
-/// Returns `Ok((Some(absolute_offset), matches))` on a successful resolution
-/// (even if the subroutine produced no matches), or `Ok((None, vec![]))`
-/// when:
-/// - the context has no rule environment attached (programmatic consumers
-///   bypassing `MagicDatabase`)
-/// - the referenced name is not in the table (logged at warn level)
-///
-/// Recursion-limit propagation is handled via [`RecursionGuard`] so that a
-/// subroutine calling `use` on itself triggers `RecursionLimitExceeded`
-/// instead of a stack overflow.
-fn evaluate_use_rule(
-    rule: &MagicRule,
-    name: &str,
-    flip_endian: bool,
-    buffer: &[u8],
-    context: &mut EvaluationContext,
-) -> Result<(Option<usize>, Vec<RuleMatch>), LibmagicError> {
-    let Some(env) = context.rule_env() else {
-        // Surface the misconfiguration once per process at warn! level so
-        // it is visible in default logging, then gate subsequent hits so a
-        // magic file with many `use` directives does not flood the log.
-        // Use `Ordering::Relaxed`: the flag is an idempotent diagnostic
-        // latch, not a synchronization primitive guarding other state.
-        if USE_WITHOUT_RULE_ENV_WARNED.swap(true, Ordering::Relaxed) {
-            debug!("use directive '{name}' evaluated without a rule environment; no-op");
-        } else {
-            warn!(
-                "use directive '{name}' evaluated without a rule environment; treating as no-op (subsequent occurrences suppressed)"
-            );
-        }
-        return Ok((None, Vec::new()));
-    };
-
-    let Some(subroutine_rules) = env.name_table.get(name) else {
-        warn!("use directive references unknown name '{name}'");
-        return Ok((None, Vec::new()));
-    };
-    // The `name` line can carry its own description (e.g. Mach-O universal
-    // `0 name mach-o \b [`, `0 name matlab4 Matlab v4 mat-file`). GNU `file`
-    // emits it ahead of the subroutine body, attached with no separating
-    // space. Capture it here while the env borrow is live; the owned `String`
-    // lets us drop that borrow before mutating the context below. `None` for
-    // a bare `name <id>`.
-    let name_message = env.name_table.name_message(name);
-    // `NameTable::get` returns an `Arc<[MagicRule]>`, so this clone is a
-    // reference-count increment rather than a deep copy of the rule tree.
-    // The Arc is cloned here to release the immutable borrow of `context`
-    // (via `env`) before we mutably borrow the context below.
-
-    // Resolve the use-site offset under the *caller's* base, not the
-    // subroutine's -- the use rule itself is in the caller's scope.
-    let absolute_offset = offset::resolve_offset_with_base(
-        &rule.offset,
-        buffer,
-        context.last_match_end(),
-        context.base_offset(),
-    )?;
-
-    // `SubroutineScope` seeds `last_match_end` and `base_offset` with
-    // the use-site offset and restores both on drop. This is the
-    // safety net for early-return paths below -- if
-    // `RecursionGuard::enter` or the inner `evaluate_rules` returns
-    // `Err(Timeout)` / `Err(RecursionLimitExceeded)`, the `?` unwinds
-    // through the guard's `Drop` impl and the caller's context
-    // returns to its pre-use state. Without the RAII wrapper a manual
-    // save/restore pair would be bypassed on every error path.
-    // Capture both the subroutine's matches AND the terminal anchor
-    // where the subroutine left `last_match_end`. The terminal anchor
-    // is what GNU `file`-compatible inlining semantics require: sibling
-    // rules after the `use` site must resolve `&N` against the position
-    // the subroutine reached, not the use-site offset. Reading the
-    // anchor INSIDE the scope (before Drop restores the caller's value)
-    // preserves it for the caller.
-    let (subroutine_matches, terminal_anchor) = {
-        let mut scope = SubroutineScope::enter(context, absolute_offset, flip_endian);
-        let mut guard = RecursionGuard::enter(scope.context())?;
-        let matches = evaluate_rules(&subroutine_rules, buffer, guard.context())?;
-        let terminal = guard.context().last_match_end();
-        (matches, terminal)
-    };
-
-    // Prepend the `name` line's own description (if any) ahead of the body's
-    // matches, matching GNU `file`: `use mach-o` emits the mach-o subroutine's
-    // `\b [` before the per-arch body. The name line reads no bytes, so this
-    // synthetic match carries a dummy value and does NOT touch the anchor
-    // (`terminal_anchor` still comes from the body's evaluation). To reproduce
-    // `file`'s no-separator attachment (`ParentSUBMSG`, not `Parent SUBMSG`),
-    // ensure the message begins with the `\b` no-separator marker
-    // (`concatenate_messages` strips a leading literal `\b` / U+0008); a name
-    // message that already starts with one (mach-o's `\b [`) is left as-is so
-    // it is not double-marked.
-    let matches = match name_message {
-        Some(msg) if !msg.is_empty() => {
-            let attached = if crate::evaluator::strip_no_separator_marker(&msg).is_some() {
-                msg
-            } else {
-                format!("\\b{msg}")
-            };
-            let name_match = RuleMatch::new(
-                attached,
-                absolute_offset,
-                rule.level,
-                crate::parser::ast::Value::Uint(0),
-                rule.typ.clone(),
-                RuleMatch::calculate_confidence(rule.level),
-            );
-            let mut combined = Vec::with_capacity(subroutine_matches.len() + 1);
-            combined.push(name_match);
-            combined.extend(subroutine_matches);
-            combined
-        }
-        _ => subroutine_matches,
-    };
-
-    Ok((Some(terminal_anchor), matches))
-}
-
-/// Evaluate a pattern-bearing rule (`TypeKind::Regex` / `TypeKind::Search`).
-///
-/// `read_pattern_match` returns `Some(value)` on a successful match
-/// (possibly zero-width, e.g., `a*`) and `None` on a genuine miss; the
-/// engine translates those directly into `Equal`/`NotEqual`. Any other
-/// operator on a pattern-bearing type is a magic-file semantic bug and
-/// surfaces as [`crate::evaluator::types::TypeReadError::UnsupportedType`] -- the earlier
-/// fallthrough to `apply_operator` masked this by producing nonsense
-/// ordering comparisons against the pattern source text.
-///
-/// On a miss we return `Value::String(String::new())` as a display
-/// placeholder; the engine has already decided `matched = false` by
-/// then, so the placeholder only affects display and
-/// `bytes_consumed_with_pattern` (which re-derives the match position
-/// from the pattern, not this value).
-fn evaluate_pattern_rule(
-    rule: &MagicRule,
-    buffer: &[u8],
-    absolute_offset: usize,
-    max_string_length: usize,
-) -> Result<(bool, crate::parser::ast::Value), LibmagicError> {
-    let match_outcome = types::read_pattern_match(
-        buffer,
-        absolute_offset,
-        &rule.typ,
-        Some(&rule.value),
-        max_string_length,
-    )
-    .map_err(|e| LibmagicError::EvaluationError(e.into()))?;
-    let pattern_found = match_outcome.is_some();
-    let matched = match &rule.op {
-        crate::parser::ast::Operator::Equal => pattern_found,
-        crate::parser::ast::Operator::NotEqual => !pattern_found,
-        other => {
-            return Err(LibmagicError::EvaluationError(
-                types::TypeReadError::UnsupportedType {
-                    type_name: format!(
-                        "operator {other:?} is not supported for pattern-bearing type {:?}; only Equal (=) and NotEqual (!=) are allowed",
-                        rule.typ
-                    ),
-                }
-                .into(),
-            ));
-        }
-    };
-    let value = match_outcome.unwrap_or_else(|| crate::parser::ast::Value::String(String::new()));
-    Ok((matched, value))
-}
-
-/// Evaluate a value-based rule (all non-pattern-bearing `TypeKind` variants).
-///
-/// Reads the typed value at `absolute_offset`, coerces the rule's
-/// expected value to the target type's signedness/width (zero-copy via
-/// `Cow::Borrowed` on the hot path), and applies the operator.
-/// `BitwiseNot` needs type-aware width masking so the complement is
-/// computed at the type's natural width (e.g. byte `NOT 0x00 = 0xFF`,
-/// not `u64::MAX`).
-fn evaluate_value_rule(
-    rule: &MagicRule,
-    buffer: &[u8],
-    absolute_offset: usize,
-    max_string_length: usize,
-    flip_endian: bool,
-) -> Result<(bool, crate::parser::ast::Value), LibmagicError> {
-    // Apply the `use \^name` endian flip (issue #236) at read time, exactly
-    // as libmagic's `cvt_flip(m->type, flip)` does in `softmagic.c`. Only the
-    // typed READ needs the flipped endianness -- `bit_width()`,
-    // `coerce_value_to_type` (the literal's numeric value is endian-invariant),
-    // the relative-offset `bytes_consumed` advance, and the string-ordering
-    // display read are all endian-invariant, so they keep `rule.typ`.
-    // `flip_type_endian` is a cheap no-op clone for the common `flip == false`
-    // path.
-    let read_typ: std::borrow::Cow<'_, crate::parser::ast::TypeKind> = if flip_endian {
-        std::borrow::Cow::Owned(types::flip_type_endian(&rule.typ))
-    } else {
-        std::borrow::Cow::Borrowed(&rule.typ)
-    };
-    let read_value = types::read_typed_value_with_pattern(
-        buffer,
-        absolute_offset,
-        read_typ.as_ref(),
-        Some(&rule.value),
-        max_string_length,
-    )
-    .map_err(|e| LibmagicError::EvaluationError(e.into()))?;
-
-    // Apply any pre-comparison value transform (`type+N`/`type-N`/`type*N`/
-    // `type/N`/`type%N`/`type|N`/`type^N`). The transform runs on the read
-    // value before the comparison operator and before printf-style format
-    // substitution, so `%d` in the message renders the post-transform
-    // number. `&MASK` is *not* handled here -- it lives at the operator
-    // layer via `Operator::BitwiseAndMask`.
-    let transformed_value = match rule.value_transform {
-        None => read_value,
-        Some(t) => operators::apply_value_transform(&read_value, t)
-            .map_err(LibmagicError::EvaluationError)?,
-    };
-
-    let expected_value = types::coerce_value_to_type(&rule.value, &rule.typ);
-    let expected_ref: &crate::parser::ast::Value = expected_value.as_ref();
-
-    let matched = match &rule.op {
-        crate::parser::ast::Operator::BitwiseNot => operators::apply_bitwise_not_with_width(
-            &transformed_value,
-            expected_ref,
-            rule.typ.bit_width(),
-        ),
-        // Masked equality (`type&MASK VALUE`) must re-normalize the masked
-        // result to the type's natural width so a signed read whose high bits
-        // are cleared by the mask still compares equal to the sign-extended
-        // rule literal (e.g. the Mach-O `0 lelong&0xfffffffe 0xfeedface`
-        // rule). See `apply_bitwise_and_mask_with_width`.
-        crate::parser::ast::Operator::BitwiseAndMask(mask) => {
-            operators::apply_bitwise_and_mask_with_width(
-                *mask,
-                &transformed_value,
-                expected_ref,
-                rule.typ.bit_width(),
-            )
-        }
-        op => operators::apply_operator(op, &transformed_value, expected_ref),
-    };
-
-    // libmagic renders the FULL string field (`p->s`) for a matched string
-    // comparison, while the comparison itself is prefix-limited to
-    // `pattern.len()` (`file_strncmp` with `vallen`). The comparison above
-    // already read exactly `pattern.len()` bytes -- correct and unchanged --
-    // but for an ORDERING operator the rendered detail needs the whole field,
-    // not the compared prefix. sgml's `>15 string/t >\0 %.3s document text`
-    // is the motivating case: comparing `>\0` reads 1 byte ("1"), but the
-    // `%.3s` must render the full field ("1.0") to produce `XML 1.0 ...`.
-    // Re-read the full field for DISPLAY only; the `matched` decision above is
-    // left byte-identical, so this cannot change any match result.
-    let display_value = string_ordering_display_value(
-        rule,
-        buffer,
-        absolute_offset,
-        max_string_length,
-        transformed_value,
-    );
-    Ok((matched, display_value))
-}
-
-/// Compute the DISPLAY value for a value-rule match, decoupling it from the
-/// value used in the comparison.
-///
-/// For a `string` rule compared with an ORDERING operator (`<`/`>`/`<=`/`>=`),
-/// libmagic renders the full string field (`p->s`, read until NUL/EOF) even
-/// though the comparison is prefix-limited to `pattern.len()`. This re-reads
-/// that full field so `%s`/`%.Ns` format specifiers render the whole value
-/// rather than only the compared prefix. For every other type or operator the
-/// compared value already IS the field libmagic renders, so `compared` is
-/// returned unchanged.
-///
-/// Only `TypeKind::String` needs this: `PString` (`read_pstring`) and
-/// `String16` (`read_string16`) already read their full field independent of
-/// `pattern.len()`, and numeric types render the whole value.
-///
-/// On a display-side read error after a successful match, the compared value
-/// is returned rather than propagating -- a matched rule must not abort on a
-/// display-only read.
-fn string_ordering_display_value(
-    rule: &MagicRule,
-    buffer: &[u8],
-    absolute_offset: usize,
-    max_string_length: usize,
-    compared: crate::parser::ast::Value,
-) -> crate::parser::ast::Value {
-    use crate::parser::ast::Operator::{GreaterEqual, GreaterThan, LessEqual, LessThan};
-
-    let is_ordering = matches!(rule.op, LessThan | GreaterThan | LessEqual | GreaterEqual);
-    if is_ordering && matches!(rule.typ, TypeKind::String { .. }) {
-        match types::read_string(buffer, absolute_offset, Some(max_string_length)) {
-            Ok(full_field) => full_field,
-            // A matched rule must not abort on a display-only read (the compared
-            // prefix was already read successfully at this offset moments ago),
-            // so fall back to it -- but `debug!` first rather than swallowing the
-            // error silently, matching this file's graceful-skip logging
-            // discipline. A latent regression here (e.g. `max_string_length`
-            // disagreeing with the original read) would otherwise render the
-            // truncated prefix with no trace of why the full field was dropped.
-            Err(e) => {
-                debug!(
-                    "string_ordering_display_value: full-field read failed at offset {absolute_offset} for rule '{}': {e}; rendering compared prefix",
-                    rule.message
-                );
-                compared
-            }
-        }
-    } else {
-        compared
-    }
-}
-
-/// Logs the graceful skip of a pattern-bearing-type rule whose
-/// `TypeReadError` is one of the two narrow allowlisted skip conditions --
-/// [`types::TypeReadError::MissingPatternOperand`] or
-/// [`types::TypeReadError::RegexCompileError`] (see
-/// `TypeReadError::is_pattern_skip` / `is_regex_compile_failure`).
-///
-/// Shared by all three engine catch sites (`evaluate_children_or_warn`, the
-/// top-level dispatch match, and the inline child-recursion match) so a
-/// future rewording of the log message only needs one touch point (DRY,
-/// AGENTS.md). Split by KTD5 (fix-system-magic-regex-graceful plan): the
-/// ordinary missing-pattern case is `debug!`-logged (an expected,
-/// low-severity data condition -- e.g. the root-cause parser
-/// miscategorization this plan also fixes), while a regex compile failure
-/// (which includes the `REGEX_COMPILE_SIZE_LIMIT` CWE-1333 denial-of-service guard) is
-/// `warn!`-logged so a malicious or pathological magic file's rejection is
-/// not silently invisible, even though the rest of the file's evaluation
-/// continues (R1: no fatal abort of the whole evaluation).
-fn log_pattern_operand_skip(site_label: &str, rule_message: &str, err: &types::TypeReadError) {
-    if err.is_regex_compile_failure() {
-        warn!(
-            "Skipping {site_label} rule '{rule_message}' due to regex compile failure: {err} -- this may indicate a malicious or pathological magic file"
-        );
-    } else {
-        debug!("Skipping {site_label} rule '{rule_message}': {err}");
-    }
-}
-
-/// Whether `message` carries any usable description text.
-///
-/// A message is considered message-less (and thus does not count as
-/// "producing output") if, after trimming ASCII/Unicode whitespace and
-/// stripping a leading GNU `file` no-separator marker (see GOTCHAS S14.1),
-/// nothing remains. This covers three shapes GNU `file` magic files use
-/// for structural/gating rules that carry no description of their own:
-/// a genuinely empty message (`""`), a whitespace-only message, and a
-/// `\b`-only message (used purely to suppress a separator when appended
-/// to a sibling's text -- with nothing else to append, it contributes no
-/// content either).
-///
-/// The marker is recognized in BOTH forms -- the raw byte `U+0008` and the
-/// literal `\b` (backslash + `'b'`) -- via the shared
-/// [`crate::evaluator::strip_no_separator_marker`], so this predicate agrees
-/// with `concatenate_messages`: a message that renders to empty there (e.g.
-/// exactly `"\b"`, the literal marker) is classified message-less here and
-/// therefore cannot win the `stop_at_first_match` race and shadow a later,
-/// more specific rule that would produce real output (the S13.2 bug class).
-fn is_message_bearing(message: &str) -> bool {
-    let trimmed = message.trim_matches(|c: char| c.is_whitespace() || c == '\u{8}');
-    let stripped = crate::evaluator::strip_no_separator_marker(trimmed).unwrap_or(trimmed);
-    !stripped
-        .trim_matches(|c: char| c.is_whitespace() || c == '\u{8}')
-        .is_empty()
-}
-
-/// Whether any match in `matches[from..]` carries usable description text
-/// (see [`is_message_bearing`]).
-///
-/// Used to decide whether a top-level rule's match -- together with any
-/// descendant matches produced by its children -- should be treated as
-/// the "winning" match for `stop_at_first_match` purposes. GNU `file`
-/// magic files commonly use message-less top-level rules purely as
-/// gating conditions for child rules (for example the `c-lang` search
-/// rules that test for `#include`/`pragma`/etc. before dispatching to a
-/// message-bearing regex child); under the old all-or-nothing contract, a
-/// message-less rule matching first under `stop_at_first_match: true`
-/// would silently shadow a later, more specific rule that actually
-/// produces a description (GOTCHAS S13.2, the assembler-source-text /
-/// plain-ASCII-text blank-output bug). A rule only "wins" the race if it
-/// (or a descendant) contributes real output text; otherwise evaluation
-/// continues to the next top-level sibling.
-///
-/// Takes `from` (the length of `matches` before this rule's dispatch) and
-/// slices via `.get()` (rather than the caller indexing `matches[from..]`
-/// directly) so this is panic-free per the project's bounds-checking
-/// discipline; `from` is always `<= matches.len()` by construction (it is
-/// captured from `matches.len()` earlier in the same call), so `.get()`
-/// always returns `Some`, but the panic-free form is required regardless.
-fn has_message_bearing_match(matches: &[RuleMatch], from: usize) -> bool {
-    matches
-        .get(from..)
-        .is_some_and(|tail| tail.iter().any(|m| is_message_bearing(&m.message)))
-}
-
-/// Evaluate a rule's children under the standard recursion-guard/graceful-skip discipline.
-///
-/// This helper centralises the `RecursionGuard` + `evaluate_rules` + error-dispatch
-/// pattern that is identical across the `Default`, `Indirect`, `Offset`, and `Use`
-/// meta-type arms in [`evaluate_rules`]. Extracting it prevents the four copies
-/// from drifting apart during future maintenance.
-///
-/// # Behaviour
-///
-/// * If `rule.children` is empty the function is a no-op (returns `Ok(())`).
-/// * Child matches are appended to `matches` in document order.
-/// * `LibmagicError::Timeout` and `LibmagicError::EvaluationError(RecursionLimitExceeded)`
-///   propagate immediately as `Err` so the caller can bail out.
-/// * Data-dependent errors (`BufferOverrun`, `InvalidOffset`,
-///   `TypeReadError::BufferOverrun`, `TypeReadError::InvalidPStringLength`,
-///   `IoError`) are logged at `warn!` and swallowed; the parent match
-///   already in `matches` is left intact. This mirrors the defensive
-///   comment in each arm: the inner `evaluate_rules` already catches and
-///   logs individual child failures, so this arm only fires if that
-///   strategy changes.
-///
-/// # Arguments
-///
-/// * `rule`      – The parent rule whose children will be evaluated.
-/// * `rule_kind` – A short label for the rule kind used in the `warn!`
-///   message (e.g. `"default"`, `"indirect"`, `"offset"`, `"use"`).
-/// * `buffer`    – The file buffer passed to the recursive call.
-/// * `context`   – Mutable evaluation context; the recursion depth is
-///   incremented on entry and decremented on drop via [`RecursionGuard`].
-/// * `matches`   – Output vector; child matches are appended here.
-fn evaluate_children_or_warn(
-    rule: &MagicRule,
-    rule_kind: &str,
-    buffer: &[u8],
-    context: &mut EvaluationContext,
-    matches: &mut Vec<RuleMatch>,
-) -> Result<(), LibmagicError> {
-    if rule.children.is_empty() {
-        return Ok(());
-    }
-    let mut guard = RecursionGuard::enter(context)?;
-    match evaluate_rules(&rule.children, buffer, guard.context()) {
-        Ok(child_matches) => {
-            matches.extend(child_matches);
-        }
-        Err(LibmagicError::Timeout { timeout_ms }) => {
-            return Err(LibmagicError::Timeout { timeout_ms });
-        }
-        // `RecursionLimitExceeded` is listed explicitly (rather than
-        // relying on the catch-all below) so a future maintainer adding
-        // another swallowed variant cannot accidentally swallow it.
-        // Both this arm and the catch-all intentionally propagate via
-        // `return Err(e)`; `match_same_arms` is suppressed because the
-        // explicit arm's purpose is documentation and future-proofing,
-        // not different behavior. See GOTCHAS S13 for the recursion-
-        // depth guard contract.
-        #[allow(clippy::match_same_arms)]
-        Err(
-            e @ LibmagicError::EvaluationError(
-                crate::error::EvaluationError::RecursionLimitExceeded { .. },
-            ),
-        ) => return Err(e),
-        Err(
-            e @ (LibmagicError::EvaluationError(
-                crate::error::EvaluationError::BufferOverrun { .. }
-                | crate::error::EvaluationError::InvalidOffset { .. }
-                | crate::error::EvaluationError::InvalidValueTransform { .. }
-                | crate::error::EvaluationError::TypeReadError(
-                    crate::evaluator::types::TypeReadError::BufferOverrun { .. }
-                    | crate::evaluator::types::TypeReadError::InvalidPStringLength { .. },
-                ),
-            )
-            | LibmagicError::IoError(_)),
-        ) => {
-            warn!(
-                "Discarding child evaluation under {} rule '{}' due to unexpected error: {} -- parent match is still emitted",
-                rule_kind, rule.message, e
-            );
-        }
-        // Narrow graceful-skip (KTD4): a pattern-bearing type evaluated
-        // without a usable pattern operand, or a regex compile failure
-        // (including the REGEX_COMPILE_SIZE_LIMIT DoS guard), must not
-        // abort the whole evaluation -- see `log_pattern_operand_skip` and
-        // the top-level dispatch match below for the full contract. This
-        // arm is defensive: under the current implementation, individual
-        // child failures are already caught and logged inside the
-        // recursive `evaluate_rules` call (they never propagate here); it
-        // guards against a future change to that strategy.
-        Err(LibmagicError::EvaluationError(crate::error::EvaluationError::TypeReadError(
-            ref tre,
-        ))) if tre.is_pattern_skip() => {
-            log_pattern_operand_skip(rule_kind, &rule.message, tre);
-        }
-        Err(e) => return Err(e),
-    }
-    Ok(())
 }
 
 /// Evaluate a list of magic rules against a file buffer with hierarchical processing
